Let release workflows run in parallel
The `release-to-github` and `release-to-npm` workflows both depend on
the `ci` workflow, but their concurrency keys collided, causing one of
the workflow runs to be cancelled.

In a reusable workflow such as `ci`, note that the `github.workflow`
context variable contains the name of the caller workflow, which is why
the `ci-` and `release-` prefixes are still required. Otherwise, the
jobs within a single release workflow would have colliding concurrency
keys.