Drop Boolean naming conventions
Usual boolean naming conventions encourage prefixes such as `is` and
`has`. However, prefixing just for the sake of prefixing is Hungarian
notation, which is generally undesirable, as it may in fact lead to less
readable names.

Native HTML elements have boolean attributes such as `autofocus`,
`checked`, `default`, and `open`, making it difficult to define a
consistent naming convention.