Prepare restoration of npm packages
With the recently added support for external configuration files, it has
become feasible to define presets for Biome. Hence, this repository
returns to define a set of npm packages with tooling presets planned for
TypeScript and Biome, undoing the original intent of commit f7c719d.
However, for maintenance reasons, I don't intend to restore the packages
for ESLint and Prettier.

I'll still be making a starter template for new web projects here:
https://github.com/spdiswal/starter-web

```shell
pnpm remove \
  tslib \
  @types/node \
  @vitest/coverage-v8 \
  @vitest/ui \
  typescript \
  vitest
```